refactor(ui): extract todo feature initializer from app.js
Move todo-domain functions into client/features/todos/initTodosFeature.js:
- handleTodoKeyPress, setPriority, getPriorityIcon
- toggleNotesInput, toggleNotes
- renderSubtasks, toggleSubtask
- aiBreakdownTodo

The feature initializer registers its own window bridge and hooks.
app.js calls initTodosFeature() during init.

app.js: 1,699 → 1,544 lines (155 lines extracted in this PR)

Closes #387
