+++ client/app.js
@@ -452,6 +452,7 @@ import {
   bindDeclarativeHandlers,
   registerServiceWorker,
 } from "./bootstrap/initGlobalListeners.js";
+import { initTodosFeature } from "./features/todos/initTodosFeature.js";
 import {
   initOnboarding,
   isOnboardingActive,
@@ -469,12 +470,8 @@ const API_URL =
     ? "http://localhost:3000"
     : window.location.origin;
 
-// Debounce utility — no external dependency.
-// Fires on the leading edge (immediate on first call) and again on the
-// trailing edge (ms after the last call), so single-event triggers (e.g.
-// Playwright fill()) respond instantly while rapid keystrokes are batched.
-const FILTER_INPUT_DEBOUNCE_MS = 180;
-// debounce + DEBOUNCED_INPUT_EXPRESSIONS moved to bootstrap/initGlobalListeners.js
+// debounce, DEBOUNCED_INPUT_EXPRESSIONS, FILTER_INPUT_DEBOUNCE_MS
+// moved to bootstrap/initGlobalListeners.js
 
 // ---------------------------------------------------------------------------
 // Module consumption — extracted pure-function modules loaded before app.js
@@ -775,171 +772,9 @@ function openTodoComposerFromCta() {
   input.focus();
 }
 
-function handleTodoKeyPress(event) {
-  if (event.key === "Enter") {
-    event.preventDefault();
-    addTodo();
-  }
-}
-
-function setPriority(priority) {
-  state.currentPriority = priority;
-
-  // Update button states
-  document.querySelectorAll(".priority-btn").forEach((btn) => {
-    btn.classList.remove("active");
-  });
-  document
-    .getElementById(
-      `priority${priority.charAt(0).toUpperCase() + priority.slice(1)}`,
-    )
-    .classList.add("active");
-  updateQuickEntryPropertiesSummary();
-}
-
-function getPriorityIcon(priority) {
-  const icons = {
-    high: "🔴",
-    medium: "🟡",
-    low: "🟢",
-  };
-  return icons[priority] || icons.medium;
-}
-
-function toggleNotesInput() {
-  const notesInput = document.getElementById("todoNotesInput");
-  const icon = document.getElementById("notesExpandIcon");
-
-  if (notesInput.style.display === "none") {
-    notesInput.style.display = "block";
-    icon.classList.add("expanded");
-  } else {
-    notesInput.style.display = "none";
-    icon.classList.remove("expanded");
-  }
-}
-
-function toggleNotes(todoId, event) {
-  event.stopPropagation();
-  const content = document.getElementById(`notes-content-${todoId}`);
-  const icon = document.getElementById(`notes-icon-${todoId}`);
-
-  if (content.style.display === "none") {
-    content.style.display = "block";
-    icon.classList.add("expanded");
-  } else {
-    content.style.display = "none";
-    icon.classList.remove("expanded");
-  }
-}
-
-function renderSubtasks(todo) {
-  const completedCount = todo.subtasks.filter((s) => s.completed).length;
-  const totalCount = todo.subtasks.length;
-
-  return `
-                <div class="subtasks-section">
-                    <div style="display: flex; justify-content: space-between; align-items: center; margin-top: 8px;">
-                        <span style="font-size: 0.85em; color: var(--text-secondary);">
-                            ☑️ Subtasks: ${completedCount}/${totalCount}
-                        </span>
-                    </div>
-                    <ul class="subtask-list">
-                        ${todo.subtasks
-                          .map(
-                            (subtask) => `
-                            <li class="subtask-item ${subtask.completed ? "completed" : ""}">
-                                <input
-                                    type="checkbox"
-                                    class="todo-checkbox"
-                                    aria-label="Mark subtask ${escapeHtml(subtask.title)} complete"
-                                    style="width: 16px; height: 16px;"
-                                    ${subtask.completed ? "checked" : ""}
-                                    data-onchange="toggleSubtask('${todo.id}', '${subtask.id}')"
-                                >
-                                <span class="subtask-title">${escapeHtml(subtask.title)}</span>
-                            </li>
-                        `,
-                          )
-                          .join("")}
-                    </ul>
-                </div>
-            `;
-}
-
-async function toggleSubtask(todoId, subtaskId) {
-  const todo = state.todos.find((t) => t.id === todoId);
-  if (!todo || !todo.subtasks) return;
-
-  const subtask = todo.subtasks.find((s) => s.id === subtaskId);
-  if (!subtask) return;
-
-  try {
-    const response = await apiCall(
-      `${API_URL}/todos/${todoId}/subtasks/${subtaskId}`,
-      {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ completed: !subtask.completed }),
-      },
-    );
-
-    if (response && response.ok) {
-      const updatedSubtask = await response.json();
-      todo.subtasks = todo.subtasks.map((s) =>
-        s.id === subtaskId ? updatedSubtask : s,
-      );
-      renderTodos();
-    }
-  } catch (error) {
-    console.error("Toggle subtask failed:", error);
-  }
-}
-
-async function aiBreakdownTodo(todoId, force = false) {
-  const todo = state.todos.find((item) => item.id === todoId);
-  if (!todo) return;
-
-  try {
-    const response = await apiCall(`${API_URL}/ai/todos/${todoId}/breakdown`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ maxSubtasks: 5, force }),
-    });
-    const data = response ? await parseApiBody(response) : {};
-
-    if (response && response.ok) {
-      await loadTodos();
-      await loadAiInsights();
-      await loadAiFeedbackSummary();
-      showMessage(
-        "todosMessage",
-        `Added ${data.createdCount || 0} AI subtasks for "${todo.title}"`,
-        "success",
-      );
-      return;
-    }
-
-    if (response && response.status === 409) {
-      const proceed = await showConfirmDialog(
-        "This task already has subtasks. Generate additional subtasks anyway?",
-      );
-      if (proceed) {
-        await aiBreakdownTodo(todoId, true);
-      }
-      return;
-    }
-
-    showMessage(
-      "todosMessage",
-      data.error || "Failed to generate subtasks",
-      "error",
-    );
-  } catch (error) {
-    console.error("AI breakdown error:", error);
-    showMessage("todosMessage", "Failed to generate subtasks", "error");
-  }
-}
+// handleTodoKeyPress, setPriority, getPriorityIcon, toggleNotesInput,
+// toggleNotes, renderSubtasks, toggleSubtask, aiBreakdownTodo
+// moved to features/todos/initTodosFeature.js
 
 document.addEventListener("keydown", function (e) {
   const isCommandK =
@@ -1333,6 +1168,8 @@ function bindDockHandlers() {
   hooks.applyTodoPatch = applyTodoPatch;
   hooks.deleteTodo = deleteTodo;
   hooks.loadTodos = loadTodos;
+  hooks.addTodo = addTodo;
+  hooks.renderTodos = renderTodos;
   hooks.validateTodoTitle = validateTodoTitle;
   hooks.toDateInputValue = toDateInputValue;
   hooks.toIsoFromDateInput = toIsoFromDateInput;
@@ -1463,13 +1300,13 @@ function bindDockHandlers() {
   hooks.processQuickEntryNaturalDate = processQuickEntryNaturalDate;
   hooks.renderOnCreateAssistRow = OnCreateAssist.renderOnCreateAssistRow;
   hooks.renderProjectOptions = renderProjectOptions;
-  hooks.renderSubtasks = renderSubtasks;
+  // hooks.renderSubtasks — set by initTodosFeature
   hooks.getProjectRecordByName = getProjectRecordByName;
   hooks.renderTodoChips = TaskDrawerAssist.renderTodoChips;
   hooks.resetQuickEntryNaturalDueState = resetQuickEntryNaturalDueState;
   hooks.selectProjectFromRail = selectProjectFromRail;
   hooks.selectWorkspaceView = selectWorkspaceView;
-  hooks.setPriority = setPriority;
+  // hooks.setPriority — set by initTodosFeature
   hooks.setQuickEntryPropertiesOpen = setQuickEntryPropertiesOpen;
   hooks.setSelectedProjectKey = setSelectedProjectKey;
   hooks.showInputDialog = showInputDialog;
@@ -1504,7 +1341,7 @@ window.addTodo = addTodo;
 window.filterTodos = filterTodos;
 window.clearFilters = clearFilters;
 window.setDateView = setDateView;
-window.handleTodoKeyPress = handleTodoKeyPress;
+// handleTodoKeyPress — registered by initTodosFeature
 window.exportVisibleTodosToIcs = exportVisibleTodosToIcs;
 // Edit modal
 window.saveEditedTodo = saveEditedTodo;
@@ -1518,8 +1355,8 @@ window.performUndo = performUndo;
 window.openTaskComposer = openTaskComposer;
 window.closeTaskComposer = closeTaskComposer;
 window.cancelTaskComposer = cancelTaskComposer;
-window.toggleNotesInput = toggleNotesInput;
-window.setPriority = setPriority;
+// toggleNotesInput — registered by initTodosFeature
+// setPriority — registered by initTodosFeature
 window.clearTaskComposerDueDate = clearTaskComposerDueDate;
 // Projects
 window.createProject = createProject;
@@ -1554,9 +1391,9 @@ window.syncSheetSearch = syncSheetSearch;
 // Todo interactions (from dynamically-rendered HTML)
 window.retryLoadTodos = retryLoadTodos;
 window.toggleTodo = toggleTodo;
-window.toggleNotes = toggleNotes;
+// toggleNotes — registered by initTodosFeature
 window.toggleSelectTodo = toggleSelectTodo;
-window.toggleSubtask = toggleSubtask;
+// toggleSubtask — registered by initTodosFeature
 window.toggleTodoKebab = toggleTodoKebab;
 window.openTodoFromKebab = openTodoFromKebab;
 window.openEditTodoFromKebab = openEditTodoFromKebab;
@@ -1579,7 +1416,7 @@ window.applyCritiqueSuggestion = applyCritiqueSuggestion;
 window.applyCritiqueSuggestionMode = applyCritiqueSuggestionMode;
 window.dismissCritiqueSuggestion = dismissCritiqueSuggestion;
 window.setCritiqueFeedbackReason = setCritiqueFeedbackReason;
-window.aiBreakdownTodo = aiBreakdownTodo;
+// aiBreakdownTodo — registered by initTodosFeature
 window.dismissPlanSuggestion = dismissPlanSuggestion;
 window.resetPlanDraft = resetPlanDraft;
 window.addPlanTasksToTodos = addPlanTasksToTodos;
@@ -1633,6 +1470,7 @@ window.runAdminFeedbackAutomation = runAdminFeedbackAutomation;
 // App bootstrap
 // ---------------------------------------------------------------------------
 function init() {
+  initTodosFeature();
   bindResponsiveLayoutState();
   renderSidebarNavigation();
   bindCriticalHandlers();
